% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate a session}
\usage{
authenticate(permission = c("read", "write", "delete"), force = FALSE,
  key = the$key, secret = the$secret)
}
\arguments{
\item{permission}{The desired permission for the session.}

\item{force}{Force requesting a new token, the default is to use the cached
token if available.}

\item{key}{The API key, a default one is provided, but please register a new
one yourself if you will be doing many requests.}

\item{secret}{The API secret.}
}
\description{
This will spawn a browser window to allow the user to accept the
authentication.
}

