% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{flickr_GET}
\alias{flickr_GET}
\alias{flickr_POST}
\title{Perform a flickr request}
\usage{
flickr_GET(method, query, ...)

flickr_POST(method, query, ...)
}
\arguments{
\item{method}{The flickr API method to use \url{https://www.flickr.com/services/api/}.}

\item{query}{list of query parameters to use in the request.}

\item{...}{additional arguments passed to \code{\link[httr]{VERB}}.}
}
\description{
\code{flickr_GET} performs a GET request, \code{flickr_POST} performs a POST
request. Use \code{\link{authenticate}} to authenticate your session prior
to connecting.
}

